requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.2.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
Flask==1.1.1
gast==0.2.2
google-auth==1.7.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.24.3
gunicorn==20.0.0
h5py==2.10.0
idna==2.8
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.2
oauthlib==3.1.0
opencv-python==4.1.1.26
opt-einsum==3.1.0
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pylint==2.3.1
pyparsing==2.4.4
python-dateutil==2.8.1
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
typed-ast==1.3.1
urllib3==1.25.6
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.1
